# msps/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "msps",
        "title": "msps",
        "contact": {},
        "version": "latest"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/a/h": {
            "post": {
                "description": "处理来自Agent的心跳请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "健康检查",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"msg\":\"\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "{\"success\":false,\"msg\":\"请求参数错误\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "{\"success\":false,\"msg\":\"用户未登录\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "403": {
                        "description": "{\"success\":false,\"msg\":\"访问受限\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "{\"success\":false,\"msg\":\"路径不存在\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "{\"success\":false,\"msg\":\"Internal Server Error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/a/m": {
            "post": {
                "description": "处理来自Agent的邮件获取请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "邮件获取",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"msg\":\"\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "{\"success\":false,\"msg\":\"请求参数错误\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "{\"success\":false,\"msg\":\"用户未登录\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "403": {
                        "description": "{\"success\":false,\"msg\":\"访问受限\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "{\"success\":false,\"msg\":\"路径不存在\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "{\"success\":false,\"msg\":\"Internal Server Error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/a/v": {
            "post": {
                "description": "处理来自Agent的邮件结果请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "邮件发送结果接收",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.EmailVerifyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"msg\":\"\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "{\"success\":false,\"msg\":\"请求参数错误\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "401": {
                        "description": "{\"success\":false,\"msg\":\"用户未登录\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "403": {
                        "description": "{\"success\":false,\"msg\":\"访问受限\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "{\"success\":false,\"msg\":\"路径不存在\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "{\"success\":false,\"msg\":\"Internal Server Error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "请求结果描述",
                    "type": "string"
                },
                "payload": {
                    "description": "请求结果数据"
                },
                "success": {
                    "description": "是否请求成功",
                    "type": "boolean"
                }
            }
        },
...
        "domain.EmailVerifyReq": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "邮件唯一标识",
                    "type": "string"
                },
                "success": {
                    "description": "邮件发送是否成功",
                    "type": "boolean"
                }
            }
        }
    }
}